+++ src/stream.rs
@@ -78,6 +78,15 @@ impl Stream {
     }
 }
 
+impl std::fmt::Debug for Stream {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Stream")
+            .field("format", &self.inner.format())
+            .field("stopped", &self.stopped)
+            .finish_non_exhaustive()
+    }
+}
+
 impl Drop for Stream {
     fn drop(&mut self) {
         if !self.stopped {